requirements.txt
# Core NLP
nltk==3.8.1
regex==2023.10.3

# Data Processing
pandas==2.1.4
numpy==1.26.4

# ML & Clustering
scikit-learn==1.4.0
scipy==1.12.0

# PDF/DOCX Parsing
PyPDF2==3.0.1
python-docx==1.1.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0
wordcloud==1.9.3

# Kaggle Dataset Download
kaggle==1.6.4

# Reporting
tabulate==0.9.0
jinja2==3.1.3

# Jupyter Notebook
notebook==7.1.0
ipykernel==6.29.2

# Environment & Utilities
python-dotenv==1.0.1
tqdm==4.66.2

# Testing
pytest==8.0.2
